pyproject.toml
[project]
name = "databot-ai"
version = "0.1.0"
description = "A lightweight, open-source AI assistant for data platform operations"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Atul Singh Bisht", email = "me.atulbisht@gmail.com"},
]
keywords = ["ai", "agent", "data", "platform", "bot", "airflow", "data-quality", "llm", "assistant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
readme = "README.md"
dependencies = [
    "typer>=0.9.0",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "httpx>=0.25.0",
    "loguru>=0.7.0",
    "rich>=13.0.0",
    "croniter>=2.0.0",
    "pyyaml>=6.0.0",
    "aiosqlite>=0.19.0",
]

[project.urls]
Homepage = "https://github.com/asb108/databot"
Repository = "https://github.com/asb108/databot"
Documentation = "https://github.com/asb108/databot#readme"
Issues = "https://github.com/asb108/databot/issues"

[project.optional-dependencies]
gchat = [
    "google-auth>=2.0.0",
    "google-api-python-client>=2.0.0",
]
sql = [
    "sqlalchemy>=2.0.0",
]
lineage = [
    "networkx>=3.0.0",
]
spark = [
    "httpx>=0.25.0",
]
kafka = [
    "httpx>=0.25.0",
]
catalog = [
    "httpx>=0.25.0",
]
glue = [
    "boto3>=1.28.0",
]
streaming = [
    "sse-starlette>=1.6.0",
]
observability = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]
rag = [
    "chromadb>=0.4.0",
]
mcp = [
    "mcp>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]
all = ["databot[gchat,sql,lineage,spark,kafka,catalog,glue,streaming,observability,rag,mcp,dev]"]

[project.scripts]
databot = "databot.cli.commands:app"

# Plugin entry points: third-party packages can register their own tools, channels, and providers
# Example: [project.entry-points."databot.tools"]
#          my_tool = "my_package:MyTool"

[project.entry-points."databot.tools"]
# Built-in tools are registered programmatically, but third-party tools use entry points

[project.entry-points."databot.channels"]
# Third-party channels (e.g., Slack, Discord) can register here

[project.entry-points."databot.providers"]
# Third-party LLM providers can register here

[project.entry-points."databot.connectors"]
sql = "databot.connectors.sql_connector:SQLConnector"
rest_api = "databot.connectors.rest_connector:RESTConnector"
catalog = "databot.connectors.catalog_connector:CatalogConnector"
spark = "databot.connectors.spark_connector:SparkConnector"
kafka = "databot.connectors.kafka_connector:KafkaConnector"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["databot"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
